+++ library/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the library will be documented in this file.
 - Add `base64` action to validate Base64 strings (pull request #644)
 - Refactor `HEXADECIMAL_REGEX` (pull request #666)
 - Change `unknown[]` in `LengthInput` type to `ArrayLike<unknown>`
+- Change `ArrayInput` and `ContentInput` type to use `MaybeReadonly`
 - Change `EMOJI_REGEX` to be more accurate and strict (pull request #666)
 - Fix bug in `fallbackAsync` method for async schemas (pull request #732)